requirements.txt
# Watch8004 - ERC-8004 AI Agent Dependencies

# Core blockchain libraries
web3==6.15.1
eth-account==0.11.0
py-solc-x==2.0.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Optional: AI/ML libraries (for enhanced capabilities)
# openai==1.12.0
# anthropic==0.18.1

# Development
pytest==8.0.0
black==24.1.1
flake8==7.0.0
